Fixes for Rust 1.66 (#851)
- Fix `clippy` warnings for unnecessary dereferences
 - Fix `clippy` warnings for unnecessary `u128 -> u128` cast
 - [`cargo publish` now blocks until it sees the published package in
   the index][1], so we can remove the retry loop from our crate
   publishing action.

[1]: https://github.com/rust-lang/cargo/blob/master/CHANGELOG.md#cargo-166-2022-12-15

+++ integration_tests/tests/common/mod.rs
@@ -270,7 +270,7 @@ pub async fn submit_measurements_and_verify_aggregate(
         VdafInstance::Prio3Aes128Sum { bits } => {
             let vdaf = Prio3::new_aes128_sum(2, *bits).unwrap();
 
-            let measurements = iter::repeat_with(|| (random::<u128>() as u128) >> (128 - bits))
+            let measurements = iter::repeat_with(|| (random::<u128>()) >> (128 - bits))
                 .take(total_measurements)
                 .collect::<Vec<_>>();
             let aggregate_result = measurements.iter().sum();
